♻️ Split api types into separate file
